Moved common methods to ExportTestHelper
